credit commit
-afimaniento creacion de cards
-integracion con eureka
