+++ pom.xml
@@ -141,9 +141,9 @@
 							<inputSpec>${project.basedir}/src/main/resources/openapi/credit-service.yaml</inputSpec>
 							<generatorName>spring</generatorName>
 							<output>${project.build.directory}/generated-sources/openapi</output>
-							<apiPackage>com.proyecto.card.api</apiPackage>
-							<modelPackage>com.proyecto.card.model</modelPackage>
-							<invokerPackage>com.proyecto.card.invoker</invokerPackage>
+							<apiPackage>com.proyecto.wallet.api</apiPackage>
+							<modelPackage>com.proyecto.wallet.model</modelPackage>
+							<invokerPackage>com.proyecto.wallet.invoker</invokerPackage>
 							<configOptions>
 								<interfaceOnly>true</interfaceOnly>
 								<useSpringBoot3>true</useSpringBoot3>
