--- src/main/java/com/proyecto/card/CreditServiceApplication.java
@@ -1,13 +1,0 @@
-package com.proyecto.card;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class CreditServiceApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(CreditServiceApplication.class, args);
-	}
-
-}
