--- src/main/java/com/proyecto/card/application/port/in/CreateCreditUseCase.java
@@ -1,8 +1,0 @@
-package com.proyecto.card.application.port.in;
-
-import com.proyecto.card.domain.model.Credit;
-import reactor.core.publisher.Mono;
-
-public interface CreateCreditUseCase {
-    Mono<Credit> create(Credit credit);
-}
