--- src/main/java/com/proyecto/card/application/port/in/DeleteCreditUseCase.java
@@ -1,7 +1,0 @@
-package com.proyecto.card.application.port.in;
-
-import reactor.core.publisher.Mono;
-
-public interface DeleteCreditUseCase {
-    Mono<Void> cancelCredit(String creditId);
-}
