--- src/main/java/com/proyecto/card/application/port/in/FindPaymentsUseCase.java
@@ -1,8 +1,0 @@
-package com.proyecto.card.application.port.in;
-
-import com.proyecto.card.domain.model.Payment;
-import reactor.core.publisher.Flux;
-
-public interface FindPaymentsUseCase {
-    Flux<Payment> findByCreditId(String creditId);
-}
