--- src/main/java/com/proyecto/card/application/port/in/GetBalanceUseCase.java
@@ -1,9 +1,0 @@
-package com.proyecto.card.application.port.in;
-
-import reactor.core.publisher.Mono;
-
-import java.math.BigDecimal;
-
-public interface GetBalanceUseCase {
-    Mono<BigDecimal> getOutstandingBalance(String creditId);
-}
