--- src/main/java/com/proyecto/card/application/port/in/RegisterPaymentUseCase.java
@@ -1,8 +1,0 @@
-package com.proyecto.card.application.port.in;
-
-import com.proyecto.card.domain.model.Payment;
-import reactor.core.publisher.Mono;
-
-public interface RegisterPaymentUseCase {
-    Mono<Payment> register(String creditId, Payment payment);
-}
