+++ src/main/java/com/proyecto/wallet/CreditServiceApplication.java
@@ -1,0 +1,13 @@
+package com.proyecto.wallet;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class CreditServiceApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(CreditServiceApplication.class, args);
+	}
+
+}
