+++ src/main/java/com/proyecto/wallet/application/port/in/CreateCreditUseCase.java
@@ -1,0 +1,8 @@
+package com.proyecto.wallet.application.port.in;
+
+import com.proyecto.wallet.domain.model.Credit;
+import reactor.core.publisher.Mono;
+
+public interface CreateCreditUseCase {
+    Mono<Credit> create(Credit credit);
+}
