+++ src/main/java/com/proyecto/wallet/application/port/in/DeleteCreditUseCase.java
@@ -1,0 +1,7 @@
+package com.proyecto.wallet.application.port.in;
+
+import reactor.core.publisher.Mono;
+
+public interface DeleteCreditUseCase {
+    Mono<Void> cancelCredit(String creditId);
+}
