+++ src/main/java/com/proyecto/wallet/application/port/in/FindPaymentsUseCase.java
@@ -1,0 +1,8 @@
+package com.proyecto.wallet.application.port.in;
+
+import com.proyecto.wallet.domain.model.Payment;
+import reactor.core.publisher.Flux;
+
+public interface FindPaymentsUseCase {
+    Flux<Payment> findByCreditId(String creditId);
+}
