+++ src/main/java/com/proyecto/wallet/application/port/in/GetBalanceUseCase.java
@@ -1,0 +1,9 @@
+package com.proyecto.wallet.application.port.in;
+
+import reactor.core.publisher.Mono;
+
+import java.math.BigDecimal;
+
+public interface GetBalanceUseCase {
+    Mono<BigDecimal> getOutstandingBalance(String creditId);
+}
