+++ src/main/java/com/proyecto/wallet/infrastructure/config/GlobalExceptionHandler.java
@@ -1,0 +1,43 @@
+package com.proyecto.wallet.infrastructure.config;
+
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+import org.springframework.web.bind.annotation.RestControllerAdvice;
+import org.springframework.web.server.ServerWebExchange;
+import reactor.core.publisher.Mono;
+
+import java.time.LocalDateTime;
+import java.util.Map;
+
+@RestControllerAdvice
+public class GlobalExceptionHandler {
+
+    @ExceptionHandler(IllegalArgumentException.class)
+    public Mono<ResponseEntity<Map<String, Object>>> handleIllegalArgument(
+            IllegalArgumentException ex, ServerWebExchange exchange) {
+
+        return Mono.just(ResponseEntity.status(HttpStatus.BAD_REQUEST)
+                .body(Map.of(
+                        "timestamp", LocalDateTime.now(),
+                        "status", HttpStatus.BAD_REQUEST.value(),
+                        "error", HttpStatus.BAD_REQUEST.getReasonPhrase(),
+                        "message", ex.getMessage(),
+                        "path", exchange.getRequest().getPath().value()
+                )));
+    }
+
+    @ExceptionHandler(Exception.class)
+    public Mono<ResponseEntity<Map<String, Object>>> handleGeneric(
+            Exception ex, ServerWebExchange exchange) {
+
+        return Mono.just(ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR)
+                .body(Map.of(
+                        "timestamp", LocalDateTime.now(),
+                        "status", HttpStatus.INTERNAL_SERVER_ERROR.value(),
+                        "error", HttpStatus.INTERNAL_SERVER_ERROR.getReasonPhrase(),
+                        "message", ex.getMessage(),
+                        "path", exchange.getRequest().getPath().value()
+                )));
+    }
+}
