+++ src/main/java/com/proyecto/wallet/infrastructure/config/OpenApiConfig.java
@@ -1,0 +1,19 @@
+package com.proyecto.wallet.infrastructure.config;
+
+import io.swagger.v3.oas.models.OpenAPI;
+import io.swagger.v3.oas.models.info.Info;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+public class OpenApiConfig {
+
+    @Bean
+    public OpenAPI customOpenAPI() {
+        return new OpenAPI()
+                .info(new Info()
+                        .title("Customer Service API")
+                        .description("REST API for managing credit in the banking system")
+                        .version("1.0.0"));
+    }
+}
