+++ src/main/resources/application.properties
@@ -7,7 +7,7 @@ spring.data.mongodb.uri = mongodb+srv://fairofox_db_user:CPs5PjoFRryUmB8h@cluste
 server.port=8082
 
 # Eureka
-eureka.client.enabled=true
+eureka.client.enabled=false
 # eureka.client.service-url.defaultZone=${EUREKA_URL:http://eureka-server:8761/eureka}
 eureka.client.service-url.defaultZone=${EUREKA_URL:http://localhost:8761/eureka}
