--- src/test/java/com/proyecto/card/CustomerServiceApplicationTests.java
@@ -1,13 +1,0 @@
-package com.proyecto.card;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class CustomerServiceApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
