+++ src/test/java/com/proyecto/wallet/CustomerServiceApplicationTests.java
@@ -1,0 +1,13 @@
+package com.proyecto.wallet;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class CustomerServiceApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
